Update
- Fixed : migrations
- Fixed : ensure product quantity appear on the POS
- Updated : ignore notifications from telescope
- Added : new queryScope on Product
- Updated : ensure all migration are run during the installation.

+++ database/migrations/v1_2/2020_10_04_093235_update_customers_addresses.php
@@ -26,7 +26,9 @@ class UpdateCustomersAddresses extends Migration
     public function down()
     {
         Schema::table( 'nexopos_customers_addresses', function( Blueprint $table ) {
-            $table->dropColumn( 'email' );
+            if ( Schema::hasColumn( 'nxopos_customers_addresses', 'email' ) ) {
+                $table->dropColumn( 'email' );
+            }
         });
     }
 }
