+++ database/migrations/v1_2/2020_10_04_172820_update_nexopos_products_table.php
@@ -27,8 +27,13 @@ class UpdateNexoposProductsTable extends Migration
     public function down()
     {
         Schema::table( 'nexopos_products', function( Blueprint $table ) {
-            $table->dropColumn( 'wholesale_tax_value' );
-            $table->dropColumn( 'sale_tax_value' );
+            if ( Schema::hasColumn( 'nexopos_products', 'wholesale_tax_value' ) ) {
+                $table->dropColumn( 'wholesale_tax_value' );
+            }
+
+            if ( Schema::hasColumn( 'nexopos_products', 'sale_tax_value' ) ) {
+                $table->dropColumn( 'sale_tax_value' );
+            }            
         });
     }
 }
